Conflicts:
	docs/source/reference/functions.rst

+++ tests/cupy_tests/random_tests/test_sample.py
@@ -60,8 +60,10 @@ class TestRandint2(unittest.TestCase):
     @condition.repeat(10)
     def test_within_interval(self):
         val = random.randint(0, 10, (2, 3)).get()
-        numpy.testing.assert_array_less(numpy.full((2, 3), -1), val)
-        numpy.testing.assert_array_less(val, numpy.full((2, 3), 10))
+        numpy.testing.assert_array_less(
+            numpy.full((2, 3), -1, dtype=numpy.int64), val)
+        numpy.testing.assert_array_less(
+            val, numpy.full((2, 3), 10, dtype=numpy.int64))
 
     @condition.retry(20)
     def test_lower_bound(self):
@@ -131,8 +133,10 @@ class TestRandomIntegers2(unittest.TestCase):
     @condition.repeat(10)
     def test_within_interval(self):
         val = random.random_integers(0, 10, (2, 3)).get()
-        numpy.testing.assert_array_less(numpy.full((2, 3), -1), val)
-        numpy.testing.assert_array_less(val, numpy.full((2, 3), 11))
+        numpy.testing.assert_array_less(
+            numpy.full((2, 3), -1, dtype=numpy.int64), val)
+        numpy.testing.assert_array_less(
+            val, numpy.full((2, 3), 11, dtype=numpy.int64))
 
     @condition.retry(20)
     def test_lower_bound(self):
